tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.pattern.crosscanvas.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.pattern.crosscanvas</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.pattern.crosscanvas</h1>
<p class="desc"></p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var response = await fetch('/images/green.png')
  var blob = await response.blob();
  var img = await createImageBitmap(blob);

  var pattern = new OffscreenCanvas(100, 50).getContext('2d').createPattern(img, 'no-repeat');
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = pattern;
  ctx.fillRect(0, 0, 100, 50);

  _assertPixel(canvas, 50,25, 0,255,0,255);

}, "");
</script>
